requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyascii==0.3.2
astroid==3.1.0
astunparse==1.6.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
bokeh==3.3.4
boto3==1.34.81
botocore==1.34.81
bs4==0.0.2
catalogue==2.0.10
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.0
contractions==0.1.73
CTkListbox==1.3
customtkinter==5.2.2
cvlib==0.2.7
cycler==0.10.0
cymem==2.0.8
Cython==3.0.10
darkdetect==0.8.0
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
dill==0.3.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
ffmpeg==1.4
filelock==3.13.4
fire==0.6.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.51.0
fpdf2==2.7.8
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
hachoir==3.3.0
html2text==2024.2.26
HTMLParser==0.0.2
huggingface-hub==0.22.2
idna==2.10
imageio==2.34.0
importlib_metadata==7.0.2
imutils==0.5.4
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.1.1
kiwisolver==1.4.5
langcodes==3.3.0
lapx==0.5.7
libclang==18.1.1
linkify-it-py==2.0.3
lxml==5.2.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
murmurhash==1.0.10
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.1
panel==1.3.8
param==2.0.2
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
preshed==3.0.9
progressbar==2.5
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyahocorasick==2.1.0
pybboxes==0.1.6
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pylint==3.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
pyviz_comms==3.0.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.7.1
roboflow==1.1.26
rpds-py==0.18.0
s3transfer==0.10.1
sahi==0.11.15
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
selectolax==0.3.21
sequential==1.0.0
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
Tcl==0.2
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
terminaltables==3.1.10
textblob==0.18.0.post0
textsearch==0.0.24
thinc==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tkcalendar==1.6.1
tkinterdnd2==0.3.0
tomlkit==0.12.4
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
ttkwidgets==0.13.0
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
uc-micro-py==1.0.3
ultralytics==8.1.46
uritools==4.0.2
urlextract==1.9.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.43.0
wordcloud==1.9.3
wrapt==1.16.0
xgboost==2.0.3
xyzservices==2023.10.1
zipp==3.17.0
